requirements.txt
fastapi
uvicorn
opencv-python-headless
numpy
Pillow
python-multipart
requests
ultralytics